Developed collaborattive rfq's filtering and accessing

+++ STCH/backend/app/services/collaboration_services/remove_collaborator.py
@@ -4,6 +4,7 @@ from app.models.rfqCollaborator_model import (
     RFQCollaborator
 )
 from sqlalchemy.orm import Session
+from app.models.rfq_model import RFQ
 
 def remove_collaborator_service(
 
@@ -31,6 +32,17 @@ def remove_collaborator_service(
         raise HTTPException(
             status_code=404,
             detail="Collaborator not found"
+        )
+    
+    rfq = db.query(RFQ).filter(
+    RFQ.id == rfq_id
+    ).first()
+
+    if rfq.created_by == user_id:
+
+        raise HTTPException(
+        status_code=400,
+        detail="RFQ creator cannot be removed"
         )
 
     db.delete(collaborator)
